Made Post Component, removed uneeded state
