+++ src/components/Feed.tsx
@@ -1,40 +1,38 @@
 import React, { useEffect } from "react";
-import * as sdk from "../services/sdk";
 import NewPost from "./NewPost";
 import PostList from "./PostList";
 import { useAppDispatch, useAppSelector } from "../redux/hooks";
+import { Button } from "antd";
+import { BaseFilters } from "@unfinishedlabs/sdk/dist/types/social/search";
+import * as sdk from "../services/sdk";
 import * as server from "../services/server";
 import {
   addNewFeedItems,
   addNewFeedtoMainFeed,
 } from "../redux/slices/feedSlice";
-import { BaseFilters } from "@unfinishedlabs/sdk/dist/types/social/search";
-import { Button } from "antd";
 
 const Feed: React.FC = () => {
-  const [subscriptionID, setSubscriptionID] = React.useState<string>("");
-
   const dispatch = useAppDispatch();
   const newFeed = useAppSelector((state) => state.feed.newFeed);
 
   useEffect(() => {
-    subscribe();
+    const subscriptionID = subscribe();
     dispatch(addNewFeedtoMainFeed());
-    return unsubscribe;
+    return () => unsubscribe(subscriptionID);
   }, []);
 
   const subscribe = () => {
     const filter: BaseFilters = {
       types: ["Broadcast", "Reply"],
     };
-    const subID = sdk.subscribeToFeed(filter, async (feedItems) => {
+    const subscriptionID = sdk.subscribeToFeed(filter, async (feedItems) => {
       const contentFeedItems = await server.loadContent(feedItems);
       dispatch(addNewFeedItems(contentFeedItems));
     });
-    setSubscriptionID(subID);
+    return subscriptionID;
   };
 
-  const unsubscribe = () => {
+  const unsubscribe = (subscriptionID: string) => {
     if (subscriptionID) sdk.unsubscribeToFeed(subscriptionID);
   };
