+++ src/components/PostList.tsx
@@ -6,6 +6,7 @@ import InfiniteScroll from "react-infinite-scroll-component";
 import * as sdk from "../services/sdk";
 import * as server from "../services/server";
 import { addFeedItems } from "../redux/slices/feedSlice";
+import Post from "./Post";
 
 const BLOCKS_PER_LOAD = 8;
 
@@ -43,11 +44,6 @@ const PostList: React.FC = () => {
     loadMoreFeed();
   }, []);
 
-  // Used for mocked data. Remove when no longer mocking data
-  const brokenLink = (e: any) => {
-    e.target.src = "https://picsum.photos/200/300";
-  };
-
   return (
     <>
       <InfiniteScroll
@@ -62,26 +58,11 @@ const PostList: React.FC = () => {
         }
       >
         <List
+          rowKey={(feedItem: FeedItem) => feedItem.hash}
           dataSource={feed || []}
           renderItem={(feedItem: FeedItem) => (
             <List.Item>
-              <div> Poster: {feedItem.address} </div>
-              <div> Hash: {feedItem.hash} </div>
-              {feedItem.replies.length > 0 && (
-                <div>
-                  {feedItem.replies.map((reply) => (
-                    <div>Reply From: {reply.address}</div>
-                  ))}
-                </div>
-              )}
-              {feedItem?.content && feedItem?.content?.attachment && (
-                <img
-                  src={(feedItem?.content).attachment[0].url}
-                  alt="Post"
-                  // Used for mocked data. Remove when no longer mocking data
-                  onError={brokenLink}
-                ></img>
-              )}
+              <Post feedItem={feedItem} />
             </List.Item>
           )}
         />
